requirements.txt
django==1.9.4
MySQL-python==1.2.5
ipython==4.1.2
ipdb==0.9.0
django-redis==4.3.0
gunicorn==19.4.5
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0
django-braces==1.8.1
xlrd==0.9.4
xlwt==1.0.0
xlsxwriter==0.8.4
Pillow==3.2.0
fab==1.4.2
pytz==2016.3
tzlocal==1.2.2
xpinyin
django-celery==3.1.17
rsa==3.4.2